move check_solver_and_constraints to _d_solver
