+++ alphacsc/tests/conftest.py
@@ -14,16 +14,14 @@ N_TIMES_ATOM, N_ATOMS = 10, 4
 parametrize_solver_and_constraint = pytest.mark.parametrize(
     'rank1, solver_d, uv_constraint',
     [
-        (True, 'auto', 'auto'),
-        (False, 'auto', 'auto'),
-        (False, 'fista', 'auto'),
-        (True, 'joint', 'auto'),
+        (True, 'alternate_adaptive', 'separate'),
+        (True, 'alternate', 'separate'),
         (True, 'joint', 'joint'),
         (True, 'joint', 'separate'),
-        (True, 'fista', 'auto'),
         (True, 'fista', 'joint'),
         (True, 'fista', 'separate'),
-        (True, 'alternate_adaptive', 'separate')
+        (False, 'auto', 'auto'),
+        (False, 'fista', 'auto'),
     ]
 )
