+++ alphacsc/tests/test_d_solver.py
@@ -8,12 +8,79 @@ from alphacsc.tests.conftest import N_TRIALS, N_CHANNELS, N_TIMES_ATOM, N_ATOMS
 from alphacsc.loss_and_gradient import compute_objective
 from alphacsc.utils import construct_X_multi
 from alphacsc.update_d_multi import prox_d, prox_uv
-from alphacsc._d_solver import get_solver_d
+from alphacsc._d_solver import get_solver_d, check_solver_and_constraints
 
 
 @pytest.fixture
 def D_init(rng, shape):
     return rng.randn(*shape)
+
+
+@pytest.mark.parametrize('solver_d', ['auto', 'fista'])
+@pytest.mark.parametrize('uv_constraint', ['auto'])
+def test_check_solver_and_constraints(solver_d, uv_constraint):
+    """Tests for valid values when rank1 is False."""
+
+    solver_d_, uv_constraint_ = check_solver_and_constraints(False, solver_d,
+                                                             uv_constraint)
+
+    assert solver_d_ == 'fista'
+    assert uv_constraint_ == 'auto'
+
+
+@pytest.mark.parametrize('solver_d', ['auto', 'fista'])
+@pytest.mark.parametrize('uv_constraint', ['joint', 'separate'])
+def test_check_solver_and_constraints_error(solver_d, uv_constraint):
+    """Tests for the case rank1 is False and params are not compatible."""
+
+    with pytest.raises(AssertionError,
+                       match="If rank1 is False, uv_constraint should be*"):
+
+        check_solver_and_constraints(False, solver_d, uv_constraint)
+
+
+@pytest.mark.parametrize('solver_d', ['auto', 'alternate',
+                                      'alternate_adaptive'])
+@pytest.mark.parametrize('uv_constraint', ['auto', 'separate'])
+def test_check_solver_and_constraints_rank1_alternate(solver_d, uv_constraint):
+    """Tests for valid values when solver is alternate and rank1 is True."""
+
+    solver_d_, uv_constraint_ = check_solver_and_constraints(True, solver_d,
+                                                             uv_constraint)
+
+    if solver_d == 'auto':
+        solver_d = 'alternate_adaptive'
+
+    assert solver_d_ == solver_d
+    assert uv_constraint_ == 'separate'
+
+
+@pytest.mark.parametrize('solver_d', ['joint', 'fista'])
+@pytest.mark.parametrize('uv_constraint', ['auto', 'joint', 'separate'])
+def test_check_solver_and_constraints_rank1(solver_d, uv_constraint):
+    """Tests for valid values when solver_d is either in 'joint' or 'fista' and
+    rank1 is True."""
+
+    solver_d_, uv_constraint_ = check_solver_and_constraints(True, solver_d,
+                                                             uv_constraint)
+
+    if uv_constraint == 'auto':
+        uv_constraint = 'joint'
+
+    assert solver_d_ == solver_d
+    assert uv_constraint_ == uv_constraint
+
+
+@pytest.mark.parametrize('solver_d', ['auto', 'alternate',
+                                      'alternate_adaptive'])
+@pytest.mark.parametrize('uv_constraint', ['joint'])
+def test_check_solver_and_constraints_rank1_error(solver_d, uv_constraint):
+    """Tests for the case when rank1 is True and params are not compatible.
+    """
+    with pytest.raises(AssertionError,
+                       match="solver_d=*"):
+
+        check_solver_and_constraints(True, solver_d, uv_constraint)
 
 
 @pytest.mark.parametrize('rank1, solver_d, uv_constraint', [
